# Conflicts:
#	Assets/Amanita/Editor/Scripts/FlowchartWindow/BlockContextMenuUitk.cs
#	Assets/Amanita/Editor/Scripts/FlowchartWindow/FcEmptySpacePopupWindow.cs
#	Assets/Amanita/Editor/Scripts/FlowchartWindow/FcWindowComponents/ClipboardObject.cs
#	Assets/Amanita/Editor/Scripts/FlowchartWindow/FcWindowComponents/FlowchartContextMenuManagerUitk.cs
#	Assets/Amanita/Editor/Scripts/FlowchartWindow/Metadata/AmanitaClipboard.cs
#	Assets/Amanita/Editor/Scripts/FlowchartWindow/Metadata/BlockClipboard.cs
#	Assets/Amanita/Editor/Scripts/FlowchartWindow/Metadata/BlockClipboardEntry.cs
#	Assets/Amanita/Editor/Scripts/FlowchartWindow/Rendering/BlockRendererUitk.cs
#	Assets/Amanita/Editor/Scripts/FlowchartWindow/Rendering/FcWindowGraphicsRendererUitk.cs
#	Assets/Amanita/Editor/Scripts/FlowchartWindow/UguiEventHandlers/DeleteShortcutHandler.cs
#	Assets/Amanita/Systems/VScripting/Flowchart.cs

+++ Assets/Amanita/Editor/Scripts/FlowchartWindow/FcWindowComponents/FcWindowRepaintTriggerer.cs
@@ -63,12 +63,12 @@ namespace Amanita.VScripting.EditorUtils
             owner = null;
         }
 
-        public void OnPostMultiBlockDeletion(IList<uint> blockIds)
+        public void OnPostMultiBlockDeletion(IList<short> blockIds)
         {
             TriggerRepaint();
         }
 
-        public void OnPostBlockDeletion(uint blockId)
+        public void OnPostBlockDeletion(ushort blockId)
         {
             TriggerRepaint();
         }
